layers: Move shared sync utils out of syncval
