+++ BUILD.gn
@@ -109,8 +109,6 @@ vvl_sources = [
   "layers/containers/small_vector.h",
   "layers/containers/span.h",
   "layers/containers/tls_guard.h",
-  "layers/containers/qfo_transfer.h",
-  "layers/containers/qfo_transfer.cpp",
   "layers/containers/range.h",
   "layers/containers/range_map.h",
   "layers/containers/subresource_adapter.cpp",
@@ -351,8 +349,6 @@ vvl_sources = [
   "layers/sync/sync_stats.h",
   "layers/sync/sync_submit.cpp",
   "layers/sync/sync_submit.h",
-  "layers/sync/sync_utils.cpp",
-  "layers/sync/sync_utils.h",
   "layers/sync/sync_validation.cpp",
   "layers/sync/sync_validation.h",
   "layers/thread_tracker/thread_safety_validation.cpp",
@@ -382,6 +378,8 @@ vvl_sources = [
   "layers/utils/ray_tracing_utils.h",
   "layers/utils/shader_utils.cpp",
   "layers/utils/shader_utils.h",
+  "layers/utils/sync_utils.cpp",
+  "layers/utils/sync_utils.h",
   "layers/utils/text_utils.cpp",
   "layers/utils/text_utils.h",
   "layers/utils/vk_layer_extension_utils.cpp",
