--- layers/containers/qfo_transfer.cpp
@@ -1,42 +1,0 @@
-/* Copyright (c) 2015-2025 The Khronos Group Inc.
- * Copyright (c) 2015-2025 Valve Corporation
- * Copyright (c) 2015-2025 LunarG, Inc.
- * Copyright (C) 2015-2025 Google Inc.
- * Modifications Copyright (C) 2020 Advanced Micro Devices, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "qfo_transfer.h"
-#include "utils/hash_vk_types.h"
-
-size_t QFOImageTransferBarrier::hash() const {
-    // Ignoring the layout information for the purpose of the hash, as we're interested in QFO release/acquisition w.r.t.
-    // the subresource affected, an layout transitions are current validated on another path
-    auto hc = base_hash_combiner() << subresourceRange;
-    return hc.Value();
-}
-
-bool QFOImageTransferBarrier::operator==(const QFOImageTransferBarrier &rhs) const {
-    // Ignoring layout w.r.t. equality. See comment in hash above.
-    return (static_cast<BaseType>(*this) == static_cast<BaseType>(rhs)) && (subresourceRange == rhs.subresourceRange);
-}
-
-size_t QFOBufferTransferBarrier::hash() const {
-    auto hc = base_hash_combiner() << offset << size;
-    return hc.Value();
-}
-
-bool QFOBufferTransferBarrier::operator==(const QFOBufferTransferBarrier &rhs) const {
-    return (static_cast<BaseType>(*this) == static_cast<BaseType>(rhs)) && (offset == rhs.offset) && (size == rhs.size);
-}
