--- layers/containers/qfo_transfer.h
@@ -1,129 +1,0 @@
-/* Copyright (c) 2015-2025 The Khronos Group Inc.
- * Copyright (c) 2015-2025 Valve Corporation
- * Copyright (c) 2015-2025 LunarG, Inc.
- * Copyright (C) 2015-2025 Google Inc.
- * Modifications Copyright (C) 2020 Advanced Micro Devices, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#pragma once
-#include "generated/error_location_helper.h"
-#include "containers/custom_containers.h"
-#include "sync/sync_utils.h"
-#include "utils/hash_util.h"
-
-namespace vvl {
-class CommandBuffer;
-}  // namespace vvl
-
-// Types to store queue family ownership (QFO) Transfers
-
-static inline bool IsOwnershipTransfer(const sync_utils::OwnershipTransferBarrier &barrier) {
-    return barrier.srcQueueFamilyIndex != barrier.dstQueueFamilyIndex;
-}
-
-static inline bool IsQueueFamilyExternal(const uint32_t queue_family_index) {
-    return (queue_family_index == VK_QUEUE_FAMILY_EXTERNAL) || (queue_family_index == VK_QUEUE_FAMILY_FOREIGN_EXT);
-}
-
-// Common to image and buffer memory barriers
-template <typename Handle>
-struct QFOTransferBarrierBase {
-    using HandleType = Handle;
-    Handle handle = VK_NULL_HANDLE;
-    uint32_t srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    uint32_t dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-
-    QFOTransferBarrierBase() = default;
-    QFOTransferBarrierBase(const Handle &resource_handle, uint32_t src, uint32_t dst)
-        : handle(resource_handle), srcQueueFamilyIndex(src), dstQueueFamilyIndex(dst) {}
-
-    hash_util::HashCombiner base_hash_combiner() const {
-        hash_util::HashCombiner hc;
-        hc << srcQueueFamilyIndex << dstQueueFamilyIndex << handle;
-        return hc;
-    }
-
-    bool operator==(const QFOTransferBarrierBase<Handle> &rhs) const {
-        return (srcQueueFamilyIndex == rhs.srcQueueFamilyIndex) && (dstQueueFamilyIndex == rhs.dstQueueFamilyIndex) &&
-               (handle == rhs.handle);
-    }
-};
-
-// Image barrier specific implementation
-struct QFOImageTransferBarrier : public QFOTransferBarrierBase<VkImage> {
-    using BaseType = QFOTransferBarrierBase<VkImage>;
-    using ImageBarrier = sync_utils::ImageBarrier;
-    VkImageLayout oldLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    VkImageLayout newLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    VkImageSubresourceRange subresourceRange;
-
-    QFOImageTransferBarrier() = default;
-    QFOImageTransferBarrier(const ImageBarrier &barrier)
-        : BaseType(barrier.image, barrier.srcQueueFamilyIndex, barrier.dstQueueFamilyIndex),
-          oldLayout(barrier.oldLayout),
-          newLayout(barrier.newLayout),
-          subresourceRange(barrier.subresourceRange) {}
-    size_t hash() const;
-    bool operator==(const QFOImageTransferBarrier &rhs) const;
-    static vvl::Struct BarrierName() { return vvl::Struct::VkImageMemoryBarrier; }
-};
-
-// Buffer barrier specific implementation
-struct QFOBufferTransferBarrier : public QFOTransferBarrierBase<VkBuffer> {
-    using BaseType = QFOTransferBarrierBase<VkBuffer>;
-    using BufferBarrier = sync_utils::BufferBarrier;
-    VkDeviceSize offset = 0;
-    VkDeviceSize size = 0;
-    QFOBufferTransferBarrier() = default;
-    QFOBufferTransferBarrier(const BufferBarrier &barrier)
-        : BaseType(barrier.buffer, barrier.srcQueueFamilyIndex, barrier.dstQueueFamilyIndex),
-          offset(barrier.offset),
-          size(barrier.size) {}
-    size_t hash() const;
-    bool operator==(const QFOBufferTransferBarrier &rhs) const;
-    static vvl::Struct BarrierName() { return vvl::Struct::VkBufferMemoryBarrier; }
-};
-
-template <typename TransferBarrier>
-using QFOTransferBarrierHash = hash_util::HasHashMember<TransferBarrier>;
-
-// Command buffers store the set of barriers recorded
-template <typename TransferBarrier>
-using QFOTransferBarrierSet = vvl::unordered_set<TransferBarrier, QFOTransferBarrierHash<TransferBarrier>>;
-
-template <typename TransferBarrier>
-struct QFOTransferBarrierSets {
-    QFOTransferBarrierSet<TransferBarrier> release;
-    QFOTransferBarrierSet<TransferBarrier> acquire;
-    void Reset() {
-        acquire.clear();
-        release.clear();
-    }
-};
-
-// The layer_data stores the map of pending release barriers
-template <typename TransferBarrier>
-using GlobalQFOTransferBarrierMap =
-    vvl::concurrent_unordered_map<typename TransferBarrier::HandleType, QFOTransferBarrierSet<TransferBarrier>>;
-
-// Submit queue uses the Scoreboard to track all release/acquire operations in a batch.
-template <typename TransferBarrier>
-using QFOTransferCBScoreboard =
-    vvl::unordered_map<TransferBarrier, const vvl::CommandBuffer *, QFOTransferBarrierHash<TransferBarrier>>;
-
-template <typename TransferBarrier>
-struct QFOTransferCBScoreboards {
-    QFOTransferCBScoreboard<TransferBarrier> acquire;
-    QFOTransferCBScoreboard<TransferBarrier> release;
-};
