+++ layers/core_checks/cc_render_pass.cpp
@@ -27,7 +27,6 @@
 #include "core_checks/cc_state_tracker.h"
 #include "core_validation.h"
 #include "error_message/error_location.h"
-#include "sync/sync_utils.h"
 #include "utils/convert_utils.h"
 #include "error_message/error_strings.h"
 #include "state_tracker/image_state.h"
