+++ layers/stateless/sl_render_pass.cpp
@@ -23,6 +23,7 @@
 #include "utils/convert_utils.h"
 #include "utils/math_utils.h"
 #include "utils/image_utils.h"
+#include "utils/sync_utils.h"
 
 namespace stateless {
