+++ layers/stateless/stateless_validation.h
@@ -21,7 +21,6 @@
 #include <vulkan/vulkan_core.h>
 #include <vulkan/utility/vk_struct_helper.hpp>
 #include "generated/error_location_helper.h"
-#include "sync/sync_utils.h"
 #include "chassis/validation_object.h"
 #include "generated/device_features.h"
