+++ layers/sync/sync_commandbuffer.cpp
@@ -31,6 +31,123 @@
 #include "utils/math_utils.h"
 #include "utils/text_utils.h"
 
+struct ShaderStageAccesses {
+    SyncAccessIndex sampled_read;
+    SyncAccessIndex storage_read;
+    SyncAccessIndex storage_write;
+    SyncAccessIndex uniform_read;
+    SyncAccessIndex acceleration_structure_read;
+};
+
+// TODO: generate me
+static ShaderStageAccesses GetShaderStageAccesses(VkShaderStageFlagBits shader_stage) {
+    static const vvl::unordered_map<VkShaderStageFlagBits, ShaderStageAccesses> map = {
+        // clang-format off
+        {VK_SHADER_STAGE_VERTEX_BIT, {
+            SYNC_VERTEX_SHADER_SHADER_SAMPLED_READ,
+            SYNC_VERTEX_SHADER_SHADER_STORAGE_READ,
+            SYNC_VERTEX_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_VERTEX_SHADER_UNIFORM_READ,
+            SYNC_VERTEX_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_TESSELLATION_CONTROL_BIT, {
+            SYNC_TESSELLATION_CONTROL_SHADER_SHADER_SAMPLED_READ,
+            SYNC_TESSELLATION_CONTROL_SHADER_SHADER_STORAGE_READ,
+            SYNC_TESSELLATION_CONTROL_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_TESSELLATION_CONTROL_SHADER_UNIFORM_READ,
+            SYNC_TESSELLATION_CONTROL_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_TESSELLATION_EVALUATION_BIT, {
+            SYNC_TESSELLATION_EVALUATION_SHADER_SHADER_SAMPLED_READ,
+            SYNC_TESSELLATION_EVALUATION_SHADER_SHADER_STORAGE_READ,
+            SYNC_TESSELLATION_EVALUATION_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_TESSELLATION_EVALUATION_SHADER_UNIFORM_READ,
+            SYNC_TESSELLATION_EVALUATION_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_GEOMETRY_BIT, {
+            SYNC_GEOMETRY_SHADER_SHADER_SAMPLED_READ,
+            SYNC_GEOMETRY_SHADER_SHADER_STORAGE_READ,
+            SYNC_GEOMETRY_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_GEOMETRY_SHADER_UNIFORM_READ,
+            SYNC_GEOMETRY_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_FRAGMENT_BIT, {
+            SYNC_FRAGMENT_SHADER_SHADER_SAMPLED_READ,
+            SYNC_FRAGMENT_SHADER_SHADER_STORAGE_READ,
+            SYNC_FRAGMENT_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_FRAGMENT_SHADER_UNIFORM_READ,
+            SYNC_FRAGMENT_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_COMPUTE_BIT, {
+            SYNC_COMPUTE_SHADER_SHADER_SAMPLED_READ,
+            SYNC_COMPUTE_SHADER_SHADER_STORAGE_READ,
+            SYNC_COMPUTE_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_COMPUTE_SHADER_UNIFORM_READ,
+            SYNC_COMPUTE_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_RAYGEN_BIT_KHR, {
+            SYNC_RAY_TRACING_SHADER_SHADER_SAMPLED_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_RAY_TRACING_SHADER_UNIFORM_READ,
+            SYNC_RAY_TRACING_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_ANY_HIT_BIT_KHR, {
+            SYNC_RAY_TRACING_SHADER_SHADER_SAMPLED_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_RAY_TRACING_SHADER_UNIFORM_READ,
+            SYNC_RAY_TRACING_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_CLOSEST_HIT_BIT_KHR, {
+            SYNC_RAY_TRACING_SHADER_SHADER_SAMPLED_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_RAY_TRACING_SHADER_UNIFORM_READ,
+            SYNC_RAY_TRACING_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_MISS_BIT_KHR, {
+            SYNC_RAY_TRACING_SHADER_SHADER_SAMPLED_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_RAY_TRACING_SHADER_UNIFORM_READ,
+            SYNC_RAY_TRACING_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_INTERSECTION_BIT_KHR, {
+            SYNC_RAY_TRACING_SHADER_SHADER_SAMPLED_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_RAY_TRACING_SHADER_UNIFORM_READ,
+            SYNC_RAY_TRACING_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_CALLABLE_BIT_KHR, {
+            SYNC_RAY_TRACING_SHADER_SHADER_SAMPLED_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_READ,
+            SYNC_RAY_TRACING_SHADER_SHADER_STORAGE_WRITE,
+            SYNC_RAY_TRACING_SHADER_UNIFORM_READ,
+            SYNC_RAY_TRACING_SHADER_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_TASK_BIT_EXT, {
+            SYNC_TASK_SHADER_EXT_SHADER_SAMPLED_READ,
+            SYNC_TASK_SHADER_EXT_SHADER_STORAGE_READ,
+            SYNC_TASK_SHADER_EXT_SHADER_STORAGE_WRITE,
+            SYNC_TASK_SHADER_EXT_UNIFORM_READ,
+            SYNC_TASK_SHADER_EXT_ACCELERATION_STRUCTURE_READ,
+        }},
+        {VK_SHADER_STAGE_MESH_BIT_EXT, {
+            SYNC_MESH_SHADER_EXT_SHADER_SAMPLED_READ,
+            SYNC_MESH_SHADER_EXT_SHADER_STORAGE_READ,
+            SYNC_MESH_SHADER_EXT_SHADER_STORAGE_WRITE,
+            SYNC_MESH_SHADER_EXT_UNIFORM_READ,
+            SYNC_MESH_SHADER_EXT_ACCELERATION_STRUCTURE_READ,
+        }},
+        // clang-format on
+    };
+    auto it = map.find(shader_stage);
+    assert(it != map.end());
+    return it->second;
+}
+
 SyncAccessIndex GetSyncStageAccessIndexsByDescriptorSet(VkDescriptorType descriptor_type,
                                                         const spirv::ResourceInterfaceVariable &variable,
                                                         VkShaderStageFlagBits stage_flag) {
@@ -41,7 +158,7 @@ SyncAccessIndex GetSyncStageAccessIndexsByDescriptorSet(VkDescriptorType descrip
         assert(stage_flag == VK_SHADER_STAGE_FRAGMENT_BIT);
         return SYNC_FRAGMENT_SHADER_INPUT_ATTACHMENT_READ;
     }
-    const auto stage_accesses = sync_utils::GetShaderStageAccesses(stage_flag);
+    const auto stage_accesses = GetShaderStageAccesses(stage_flag);
 
     if (descriptor_type == VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER || descriptor_type == VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC) {
         return stage_accesses.uniform_read;
